Minor: Move `PlanType`, `StringifiedPlan` and `ToStringifiedPlan`  `datafusion_common`  (#6571)
* Move DisplayablePlan to `datafusion_common`

* Update uses

* Update datafusion/common/src/display.rs


---------
